Added `CrossBuildInMemoryCacheFactory.newClassCache()` which can be used to create a cache whose keys are Class instances, implemented in an efficient way that also allows classes to be collected when no longer required.
The implementation uses the same simple strategy that the caches created by `newCache()` use, with different handling of the keys. Both cache implementations retain strong references to the keys and the values for all entries that are read or put during the current build session and the previous build session, and soft references for the values for all other entries. The difference between `newCache()` and `newClassCache()` is that `newCache()` uses strong references to all keys and `newClassCache()` uses weak references to keys for entries that have not been used recently.

This strategy is intentionally really simple, but should be good enough and can be made better later. For example, the implementations might later receive health monitoring notifications and release entries. Or they might receive notifications that a `ClassLoader` is no longer required (for example its classpath has changed) and can discard entries for classes of that `ClassLoader`.